+++ go/core/internal/service/system/service.go
@@ -12,7 +12,7 @@ import (
 	"github.com/kagent-dev/kagent/go/core/internal/service/serviceerrors"
 	"github.com/kagent-dev/kagent/go/core/internal/version"
 	"github.com/kagent-dev/kagent/go/core/pkg/auth"
-	"github.com/kagent-dev/kagent/go/core/pkg/sandboxbackend/substrate"
+	"github.com/kagent-dev/kagent/go/core/v2/substrate"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -330,8 +330,8 @@ func (s *Service) listSubstrateCRs(ctx context.Context, namespace string) ([]Sub
 			WorkerSelector:  labelSelectorString(ctx, actorTemplate.Spec.WorkerSelector),
 			ManagedByKagent: actorTemplate.Labels["app.kubernetes.io/managed-by"] == "kagent",
 		}
-		if agentName := substrate.SandboxAgentNameFromLabels(actorTemplate.Labels); agentName != "" {
-			entry.HarnessName = agentName
+		if harnessName := actorTemplate.Labels[substrate.RevisionHarnessLabel]; harnessName != "" {
+			entry.HarnessName = harnessName
 		}
 		actorTemplates = append(actorTemplates, entry)
 	}
