--- go/core/pkg/sandboxbackend/async.go
@@ -1,56 +1,0 @@
-package sandboxbackend
-
-import (
-	"context"
-
-	"github.com/kagent-dev/kagent/go/api/v1alpha3"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// Handle is the opaque identifier an AsyncBackend uses to address a sandbox
-// it owns on an external control plane. Persisted in AgentHarness.Status.BackendRef.
-//
-// For Substrate backends, Atespace scopes the ID — an actor's identity on
-// substrate is (Atespace, ID). Non-substrate backends may leave Atespace empty.
-type Handle struct {
-	ID       string
-	Atespace string
-}
-
-// EnsureResult is returned by EnsureAgentHarness. Endpoint (if set) is surfaced
-// to users via AgentHarness.Status.Connection (Substrate: kagent gateway proxy path).
-type EnsureResult struct {
-	Handle   Handle
-	Endpoint string
-}
-
-// AsyncBackend is the minimal surface a gRPC/HTTP-driven sandbox control
-// plane must implement to back the kagent.dev/v1alpha3 AgentHarness CRD. It is
-// deliberately separate from Backend (which serves SandboxAgent's in-cluster
-// agent-runtime flow).
-type AsyncBackend interface {
-	// Name identifies the backend for AgentHarness.Status.BackendRef.Backend
-	// and logging.
-	Name() v1alpha3.AgentHarnessBackendType
-
-	// EnsureAgentHarness creates the sandbox on the backend if it does not
-	// already exist. Implementations must be idempotent — if a sandbox
-	// matching sbx.Name is already present, return its current handle.
-	EnsureAgentHarness(ctx context.Context, ah *v1alpha3.AgentHarness) (EnsureResult, error)
-
-	// GetStatus returns a Ready condition (status, reason, message) for
-	// the sandbox identified by h. Used to refresh AgentHarness.Status after
-	// each reconcile.
-	GetStatus(ctx context.Context, h Handle) (metav1.ConditionStatus, string, string)
-
-	// DeleteAgentHarness releases the sandbox. It performs at most one
-	// reconcile-safe delete step and returns done=true once the sandbox is gone.
-	// NotFound must be treated as success so the finalizer can be removed
-	// idempotently.
-	DeleteAgentHarness(ctx context.Context, h Handle) (done bool, err error)
-
-	// OnAgentHarnessReady runs one-time work after the AgentHarness reports
-	// Ready (for example ExecSandbox bootstrap inside the VM). Backends that
-	// have no post-ready work should return nil.
-	OnAgentHarnessReady(ctx context.Context, ah *v1alpha3.AgentHarness, h Handle) error
-}
