--- go/core/pkg/sandboxbackend/backend.go
@@ -1,36 +1,0 @@
-package sandboxbackend
-
-import (
-	"context"
-
-	"github.com/kagent-dev/kagent/go/api/v1alpha3"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-)
-
-// BuildInput carries the pod template for a Sandbox workload (agents.x-k8s.io Sandbox).
-type BuildInput struct {
-	Agent        *v1alpha3.SandboxAgent
-	PodTemplate  corev1.PodTemplateSpec
-	ConfigSecret *corev1.Secret
-	WorkloadName string
-	ExtraLabels  map[string]string
-}
-
-// Backend builds sandbox CRD objects and evaluates their readiness.
-type Backend interface {
-	BuildSandbox(ctx context.Context, in BuildInput) ([]client.Object, error)
-	GetOwnedResourceTypes() []client.Object
-	// OwnedResourceTypesFor returns owned types for the agent's sandbox platform (for reconcile lists).
-	OwnedResourceTypesFor(agent *v1alpha3.SandboxAgent) ([]client.Object, error)
-
-	// SessionDBURL returns the backend-specific session-store URL the translator bakes into the
-	// agent's rendered config (AgentConfig.session_db_url) before building the config Secret,
-	// or "" when the backend keeps sessions in the controller database.
-	SessionDBURL(agent *v1alpha3.SandboxAgent) string
-
-	// ComputeReady reflects implementation-specific status into condition pieces for Agent.status.
-	ComputeReady(ctx context.Context, cl client.Client, nn types.NamespacedName) (status metav1.ConditionStatus, reason, message string)
-}
