--- go/core/pkg/sandboxbackend/filter_translator_owned_test.go
@@ -1,52 +1,0 @@
-package sandboxbackend_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	atev1alpha1 "github.com/agent-substrate/substrate/pkg/api/v1alpha1"
-	"github.com/kagent-dev/kagent/go/api/v1alpha3"
-	"github.com/kagent-dev/kagent/go/core/pkg/sandboxbackend"
-	"github.com/kagent-dev/kagent/go/core/pkg/sandboxbackend/substrate"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-)
-
-func TestFilterTranslatorOwnedTypesForList(t *testing.T) {
-	scheme := runtime.NewScheme()
-	require.NoError(t, corev1.AddToScheme(scheme))
-	require.NoError(t, atev1alpha1.AddToScheme(scheme))
-	require.NoError(t, v1alpha3.AddToScheme(scheme))
-
-	cl := fake.NewClientBuilder().WithScheme(scheme).Build()
-	backend := substrate.NewAgentsBackend(nil, nil)
-
-	allTypes := []client.Object{
-		&corev1.ConfigMap{},
-		&atev1alpha1.ActorTemplate{},
-	}
-
-	t.Run("SandboxAgent drops ActorTemplate from prune (managed via blue-green)", func(t *testing.T) {
-		sa := &v1alpha3.SandboxAgent{ObjectMeta: metav1.ObjectMeta{Name: "s", Namespace: "ns"}}
-		out, err := sandboxbackend.FilterTranslatorOwnedTypesForList(cl, sa, allTypes, backend)
-		require.NoError(t, err)
-		// Substrate manages ActorTemplate lifecycle itself (config-hashed, blue-green), so it is
-		// excluded from the generic prune list — leaving only the ConfigMap.
-		require.Len(t, out, 1)
-		for _, o := range out {
-			_, ok := o.(*atev1alpha1.ActorTemplate)
-			require.False(t, ok, "ActorTemplate must be excluded from generic prune (managed via blue-green)")
-		}
-	})
-
-	t.Run("nil backend is passthrough", func(t *testing.T) {
-		agent := &v1alpha3.SandboxAgent{ObjectMeta: metav1.ObjectMeta{Name: "a", Namespace: "ns"}}
-		out, err := sandboxbackend.FilterTranslatorOwnedTypesForList(cl, agent, allTypes, nil)
-		require.NoError(t, err)
-		require.Len(t, out, len(allTypes))
-	})
-}
