--- go/core/pkg/sandboxbackend/substrate/actor_errors.go
@@ -1,48 +1,0 @@
-package substrate
-
-import (
-	"errors"
-	"fmt"
-	"strings"
-
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
-)
-
-// ErrNoFreeWorkers is returned when ate-api cannot assign a WorkerPool worker to resume an actor.
-var ErrNoFreeWorkers = errors.New("substrate worker pool has no free workers; try again later or increase WorkerPool replicas")
-
-func wrapResumeActorError(actorID string, err error) error {
-	if err == nil {
-		return nil
-	}
-	if isNoFreeWorkersError(err) {
-		return fmt.Errorf("%w", ErrNoFreeWorkers)
-	}
-	return fmt.Errorf("substrate ResumeActor %q: %w", actorID, err)
-}
-
-// wrapCreateActorError normalizes a CreateActor failure, surfacing the clean ErrNoFreeWorkers
-// (rather than an opaque gRPC FailedPrecondition) when the WorkerPool is at capacity so the chat
-// caller can return an actionable "no free workers" error.
-func wrapCreateActorError(actorID string, err error) error {
-	if err == nil {
-		return nil
-	}
-	if isNoFreeWorkersError(err) {
-		return fmt.Errorf("%w", ErrNoFreeWorkers)
-	}
-	return fmt.Errorf("substrate CreateActor %q: %w", actorID, err)
-}
-
-func isNoFreeWorkersError(err error) bool {
-	if errors.Is(err, ErrNoFreeWorkers) {
-		return true
-	}
-	if st, ok := status.FromError(err); ok {
-		if st.Code() == codes.FailedPrecondition && strings.Contains(strings.ToLower(st.Message()), "no free workers") {
-			return true
-		}
-	}
-	return strings.Contains(strings.ToLower(err.Error()), "no free workers")
-}
