+++ go/core/v2/substrate/gateway_test.go
@@ -1,0 +1,9 @@
+package substrate
+
+import "testing"
+
+func TestActorHost(t *testing.T) {
+	if got := ActorHost("kagent", "actor-1", ""); got != "actor-1.kagent.actors.resources.substrate.ate.dev" {
+		t.Fatalf("ActorHost() = %q", got)
+	}
+}
